+++ backend/src/handlers/poker-tables/PokerTablesLeaveHandler.test.ts
@@ -1,11 +1,10 @@
-// Internal
-import { GameLobbyService } from '../../game-lobby-service';
-import { shutDownServer } from '@tests/helpers/shutDownServer';
-import { Rooms } from '../../sockets/Rooms';
-import { ResultSuccess, ResultError } from '@infra/Result';
-import { RoomNotFoundError } from '../../sockets/errors/RoomErrors';
-
 import { apiTest } from '@tests/helpers/apiTest';
+import { shutDownServer } from '@tests/helpers/shutDownServer';
+import { ResultSuccess, ResultError } from '@infra/Result';
+
+import { GameLobbyService } from '../../game-lobby-service';
+import { Rooms } from '../../sockets/Rooms';
+import { RoomNotFoundError } from '../../sockets/errors/RoomErrors';
 
 describe('poker-tables.leave', () => {
   // TODO: need to add more unit tests for invalid requests and types
@@ -18,7 +17,6 @@ describe('poker-tables.leave', () => {
       socketId: 'abc123',
     });
 
-    expect(res.statusCode).toEqual(400);
     expect(res.body.ok).toEqual(false);
     expect(res.body.error.errorCode).toEqual('invalid_request_payload');
   });
