+++ backend/src/handlers/users/UsersCreateHandler.test.ts
@@ -1,5 +1,5 @@
-import { shutDownServer } from '@tests/helpers/shutDownServer';
 import { apiTest } from '@tests/helpers/apiTest';
+import { shutDownServer } from '@tests/helpers/shutDownServer';
 
 describe('users.create', () => {
   it('should error when payload is invalid', async () => {
@@ -7,7 +7,6 @@ describe('users.create', () => {
       username: '',
       password: 'abc123',
     });
-    expect(res.statusCode).toEqual(400);
     expect(res.body.ok).toEqual(false);
     expect(res.body.error.errorCode).toEqual('invalid_request_payload');
   });
