+++ backend/src/handlers/users/errors/UserNotAuthedError.ts
@@ -1,0 +1,7 @@
+import { BaseError } from '@infra/BaseError';
+
+export class UserNotAuthedError extends BaseError {
+  constructor() {
+    super('not_authed', 'User not authenticated');
+  }
+}
